refractor: rename class files using snake_case
