+++ w3d_translator/classes/game_object.py
