+++ w3d_translator/classes/xml_dataclass.py
